+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -37,6 +37,7 @@ import dev.zacsweers.metro.compiler.ir.finalizeFakeOverride
 import dev.zacsweers.metro.compiler.ir.getAllSuperTypes
 import dev.zacsweers.metro.compiler.ir.getConstBooleanArgumentOrNull
 import dev.zacsweers.metro.compiler.ir.getSingleConstBooleanArgumentOrNull
+import dev.zacsweers.metro.compiler.ir.hiddenDeprecated
 import dev.zacsweers.metro.compiler.ir.irExprBodySafe
 import dev.zacsweers.metro.compiler.ir.irInvoke
 import dev.zacsweers.metro.compiler.ir.irLambda
@@ -126,6 +127,7 @@ import org.jetbrains.kotlin.ir.util.callableId
 import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.companionObject
+import org.jetbrains.kotlin.ir.util.deepCopyWithSymbols
 import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.dumpKotlinLike
 import org.jetbrains.kotlin.ir.util.functions
@@ -1808,8 +1810,7 @@ internal class DependencyGraphTransformer(
               addFunction(
                   name = "${field.name.asString()}${Symbols.StringNames.METRO_ACCESSOR}",
                   returnType = field.type,
-                  // TODO is this... ok?
-                  visibility = DescriptorVisibilities.INTERNAL,
+                  visibility = DescriptorVisibilities.PUBLIC,
                   origin = Origins.InstanceFieldAccessor,
                 )
                 .apply {
@@ -1818,7 +1819,8 @@ internal class DependencyGraphTransformer(
                       it.ir.transform(DeepCopyIrTreeWithSymbols(SymbolRemapper.EMPTY), null)
                         as IrConstructorCall
                   }
-                  // TODO add deprecation + hidden annotation to hide? Not sure if necessary
+                  // Add Deprecated(HIDDEN) annotation to hide
+                  annotations += hiddenDeprecated()
                   body =
                     pluginContext.createIrBuilder(symbol).run {
                       val expression =
@@ -2522,7 +2524,7 @@ internal class DependencyGraphTransformer(
         mapKey
       } else {
         // We can just copy the expression!
-        mapKey.getValueArgument(0)!!
+        mapKey.getValueArgument(0)!!.deepCopyWithSymbols()
       }
 
     return expression
@@ -2746,7 +2748,7 @@ internal class DependencyGraphTransformer(
         val lambda =
           irLambda(
             context = pluginContext,
-            parent = generationContext.thisReceiver.parent,
+            parent = this.parent,
             receiverParameter = null,
             emptyList(),
             binding.typeKey.type,